+++ metro-ai-suite/metro-sdk-manager/docs/oep-sdk-manager-files/config.js
@@ -57,6 +57,10 @@ const CONFIG = {
           value: "latest"
         },
         {
+          label: "2026.2",
+          value: "2026.2"
+        },
+        {
           label: "2026.0",
           value: "2026.0"
         },
@@ -73,6 +77,57 @@ const CONFIG = {
       label: "Installed Components",
       fallback: "Select options to see installed components…",
       rules: [
+        {
+          when: {
+            SDK: "OEP_VISION",
+            OP_SYSTEM: "UBUNTU",
+            VERSION: "2026.2"
+          },
+          components: [
+            "DL Streamer",
+            "DL Streamer Pipeline Server",
+            "OpenVINO",
+            "OpenVINO Model Server",
+            "Scenescape Manager",
+            "Scenescape Controller",
+            "Scenescape Autocalibration",
+            "Edge AI Libraries - Repo",
+            "Edge AI Suites - Repo",
+            "Scenescape - Repo"
+          ]
+        },
+        {
+          when: {
+            SDK: "OEP_GENAI",
+            OP_SYSTEM: "UBUNTU",
+            VERSION: "2026.2"
+          },
+          components: [
+            "Audio Analyzer Microservice",
+            "Document Ingestion (pgvector)",
+            "Multimodal Embedding Serving",
+            "Multimodal Data Preparation",
+            "Model Download",
+            "Chat Q&A Core",
+            "Edge AI Libraries - Repo",
+            "Edge AI Suites - Repo"
+          ]
+        },
+        {
+          when: {
+            SDK: "VISUAL_AI_DEMO",
+            OP_SYSTEM: "UBUNTU",
+            VERSION: "2026.2"
+          },
+          components: [
+            "DL Streamer Pipeline Server",
+            "Node Red",
+            "Grafana",
+            "MediaMTX",
+            "MQTT Broker",
+            "Edge AI Suites - Repo"
+          ]
+        },
         {
           when: {
             SDK: "OEP_VISION",
@@ -247,6 +302,32 @@ const CONFIG = {
           when: {
             SDK: "OEP_VISION",
             OP_SYSTEM: "UBUNTU",
+            VERSION: "2026.2"
+          },
+          text: `curl -fsS https://raw.githubusercontent.com/open-edge-platform/edge-ai-suites/refs/heads/release-2026.2.0/metro-ai-suite/metro-sdk-manager/scripts/oep-vision-ai-sdk.sh | bash`
+        },
+
+        {
+          when: {
+            SDK: "OEP_GENAI",
+            OP_SYSTEM: "UBUNTU",
+            VERSION: "2026.2"
+          },
+          text: `curl -fsS https://raw.githubusercontent.com/open-edge-platform/edge-ai-suites/refs/heads/release-2026.2.0/metro-ai-suite/metro-sdk-manager/scripts/oep-gen-ai-sdk.sh | bash`
+        },
+
+        {
+          when: {
+            SDK: "VISUAL_AI_DEMO",
+            OP_SYSTEM: "UBUNTU",
+            VERSION: "2026.2"
+          },
+          text: `curl -fsS https://raw.githubusercontent.com/open-edge-platform/edge-ai-suites/refs/heads/release-2026.2.0/metro-ai-suite/metro-sdk-manager/scripts/visual-ai-demo-kit.sh | bash`
+        },
+        {
+          when: {
+            SDK: "OEP_VISION",
+            OP_SYSTEM: "UBUNTU",
             VERSION: "2026.1"
           },
           text: `curl -fsS https://raw.githubusercontent.com/open-edge-platform/edge-ai-suites/refs/heads/release-2026.1.0/metro-ai-suite/metro-sdk-manager/scripts/oep-vision-ai-sdk.sh | bash`
@@ -337,6 +418,33 @@ const CONFIG = {
       label: "Next Steps",
       fallback: "Select options to see next steps…",
       rules: [
+        {
+          when: {
+            SDK: "OEP_VISION",
+            OP_SYSTEM: "UBUNTU",
+            VERSION: "2026.2"
+          },
+          text: `Get Started`,
+          link: `https://docs.openedgeplatform.intel.com/2026.2/OEP-articles/oep-sdk-manager/oep-vision-ai-sdk/get-started.html`
+        },
+        {
+          when: {
+            SDK: "OEP_GENAI",
+            OP_SYSTEM: "UBUNTU",
+            VERSION: "2026.2"
+          },
+          text: `Get Started`,
+          link: `https://docs.openedgeplatform.intel.com/2026.2/OEP-articles/oep-sdk-manager/oep-gen-ai-sdk/get-started.html`
+        },
+        {
+          when: {
+            SDK: "VISUAL_AI_DEMO",
+            OP_SYSTEM: "UBUNTU",
+            VERSION: "2026.2"
+          },
+          text: `Get Started`,
+          link: `https://docs.openedgeplatform.intel.com/2026.2/OEP-articles/oep-sdk-manager/visual-ai-demo-kit/get-started.html`
+        },
         {
           when: {
             SDK: "OEP_VISION",
@@ -434,6 +542,51 @@ const CONFIG = {
       label: "Resources",
       fallback: "Select options to see resources…",
       rules: [
+        {
+          when: {
+            SDK: "OEP_VISION",
+            OP_SYSTEM: "UBUNTU",
+            VERSION: "2026.2"
+          },
+          links: [
+            { text: "DL Streamer", url: "http://docs.openedgeplatform.intel.com/2026.2/edge-ai-libraries/dl-streamer/index.html" },
+            { text: "DL Streamer Pipeline Server", url: "https://docs.openedgeplatform.intel.com/2026.2/edge-ai-libraries/dlstreamer-pipeline-server/index.html" },
+            { text: "OpenVINO", url: "https://docs.openvino.ai/2026/get-started.html" },
+            { text: "OpenVINO Model Server", url: "https://docs.openvino.ai/2026/model-server/ovms_what_is_openvino_model_server.html" },
+            { text: "Scenescape", url: "https://github.com/open-edge-platform/scenescape" },
+            { text: "Edge AI Libraries", url: "https://docs.openedgeplatform.intel.com/2026.2/ai-libraries.html"},
+            { text: "Edge AI Suites", url: "https://docs.openedgeplatform.intel.com/2026.2/ai-suite-metro.html"}
+          ]
+        },
+        {
+          when: {
+            SDK: "OEP_GENAI",
+            OP_SYSTEM: "UBUNTU",
+            VERSION: "2026.2"
+          },
+          links: [
+            { text: "Audio Analyzer", url: "https://docs.openedgeplatform.intel.com/2026.2/edge-ai-libraries/audio-analyzer/index.html" },
+            { text: "Document Ingestion - pgvector", url: "https://docs.openedgeplatform.intel.com/2026.2/edge-ai-libraries/pgvector/index.html" },
+            { text: "Multimodal Embedding Serving", url: "https://docs.openedgeplatform.intel.com/2026.2/edge-ai-libraries/multimodal-embedding-serving/index.html" },
+            { text: "Multimodal Data Preparation", url: "https://github.com/open-edge-platform/edge-ai-libraries/blob/release-2026.2.0/microservices/visual-data-preparation-for-retrieval/multimodal-dataprep/docs/user-guide/Overview.md" },
+            { text: "Chat Q&A Core", url: "http://docs.openedgeplatform.intel.com/2026.2/edge-ai-libraries/chat-question-and-answer-core/index.html" },
+            { text: "Edge AI Libraries", url: "https://docs.openedgeplatform.intel.com/2026.2/ai-libraries.html"},
+            { text: "Edge AI Suites", url: "https://docs.openedgeplatform.intel.com/2026.2/ai-suite-metro.html"}
+          ]
+        },
+        {
+          when: {
+            SDK: "VISUAL_AI_DEMO",
+            OP_SYSTEM: "UBUNTU",
+            VERSION: "2026.2"
+          },
+          links: [
+            { text: "DL Streamer", url: "http://docs.openedgeplatform.intel.com/2026.2/edge-ai-libraries/dl-streamer/index.html" },
+            { text: "DL Streamer Pipeline Server", url: "https://docs.openedgeplatform.intel.com/2026.2/edge-ai-libraries/dlstreamer-pipeline-server/index.html" },
+            { text: "Edge AI Libraries", url: "https://docs.openedgeplatform.intel.com/2026.2/ai-libraries.html"},
+            { text: "Edge AI Suites", url: "https://docs.openedgeplatform.intel.com/2026.2/ai-suite-metro.html"}
+          ]
+        },
         {
           when: {
             SDK: "OEP_VISION",