+++ metro-ai-suite/metro-sdk-manager/scripts/visual-ai-demo-kit.sh
@@ -18,16 +18,16 @@
 
 # Array of repositories to clone: "url|branch|directory"
 repositories=(
-  "https://github.com/open-edge-platform/edge-ai-suites|release-2026.1.0|edge-ai-suites"
+  "https://github.com/open-edge-platform/edge-ai-suites|release-2026.2.0|edge-ai-suites"
 )
 
 images=(
   eclipse-mosquitto:2.0.22
-  intel/dlstreamer-pipeline-server:2026.1.0-ubuntu24
-  bluenviron/mediamtx:1.19.1
-  coturn/coturn:4.13.1
-  grafana/grafana:13.0.2
-  nodered/node-red:5.0.0
+  intel/dlstreamer-pipeline-server:2026.2.0-ubuntu24-rc1
+  bluenviron/mediamtx:1.20.0
+  coturn/coturn:4.17.0
+  grafana/grafana:13.1.3
+  nodered/node-red:5.0.4
 )
 
 NAME="Visual AI Demo Kit"